UnitTestFrameworkPkg: Update ReadMe.md based on review findings
Fix FunctionMockLib example of the MockUefiLib declaration to align with the
design code.

Fix FunctionMockLib Mocks header file location paths in the tables.

Fix FunctionMockLib Mocks example (and description) of the MockUefiLib
declaration to align with the design code.

Fix FunctionMockLib Mocks example (and description) of the MockUefiLib INF
file to align with the design file.

Fix typos in new Code Coverage section.

Signed-off-by: Chris Johnson <chris.n.johnson@intel.com>